refactor: Wordle 패키지 이동
